requirements.txt
griptape
python-dotenv
click